+++ activity_browser/actions/__init__.py
@@ -11,6 +11,7 @@ from .activity.activity_new_product import ActivityNewProduct
 from .activity.activity_open import ActivityOpen
 from .activity.activity_relink import ActivityRelink
 from .activity.node_properties import NodeProperties
+from .activity.activity_redo_allocation import MultifunctionalProcessRedoAllocation
 from .biosphere_update import BiosphereUpdate
 
 from .calculation_setup.cs_new import CSNew
