+++ activity_browser/layouts/pages/__init__.py
@@ -1,0 +1,1 @@
+from .activity_details import ActivityDetails
