+++ activity_browser/layouts/tabs/activity.py
@@ -1,33 +1,14 @@
 # -*- coding: utf-8 -*-
 from logging import getLogger
 
-from bw2data import databases, get_node
-from bw2data.proxies import ExchangeProxyBase
-from peewee import DoesNotExist
-from qtpy import QtCore, QtWidgets
-from qtpy.QtCore import Slot, Qt
+from qtpy.QtCore import Slot
 
-from activity_browser import ab_settings, project_settings, signals
-from activity_browser.actions.activity.activity_redo_allocation import MultifunctionalProcessRedoAllocation
+import bw2data as bd
+
+from activity_browser import signals
 from activity_browser.bwutils import commontasks as bc
-from activity_browser.mod import bw2data as bd
-from activity_browser.ui.widgets.custom_allocation_editor import CustomAllocationEditor
-from activity_browser.ui.widgets.property_editor import PropertyEditor
+from activity_browser.layouts.pages import ActivityDetails
 
-from ...ui.icons import qicons
-from ...ui.style import style_activity_tab
-from ...ui.tables import (
-    BiosphereExchangeTable,
-    DownstreamExchangeTable,
-    ProductExchangeTable,
-    TechnosphereExchangeTable,
-)
-from ...ui.widgets import (
-    ActivityDataGrid,
-    DetailsGroupBox,
-    SignalledPlainTextEdit,
-    TagEditor,
-)
 from ..panels.panel import ABTab
 
 log = getLogger(__name__)
@@ -53,13 +34,7 @@ class ActivitiesTab(ABTab):
         """Opens new tab or focuses on already open one."""
         if key not in self.tabs:
             act = bd.get_activity(key)
-            # if act._document.type not in bd.labels.node_types:
-            #     QtWidgets.QMessageBox.information(self,
-            #         "Node can't be displayed",
-            #         "Node can't be displayed because it isn't a process or a product",
-            #         QtWidgets.QMessageBox.StandardButton.Ok)
-            #     return
-            new_tab = ActivityTab(key, read_only, self)
+            new_tab = ActivityDetails(key, read_only, self)
 
             # If this is a new or duplicated activity then we want to exit it
             # ditto check the Technosphere and Biosphere tables
@@ -89,314 +64,3 @@ class ActivitiesTab(ABTab):
     def safe_open_activity_tab(self, key: tuple) -> None:
         self.open_activity_tab(key)
 
-
-class ActivityTab(QtWidgets.QWidget):
-    """The data relating to Brightway activities can be viewed and edited through this panel interface
-    The interface is a GUI representation of the standard activity data format as determined by Brightway
-    This is necessitated as AB does not save its own data structures to disk
-    Data format documentation is under the heading "The schema for an LCI dataset in voluptuous is:" at this link:
-    https://docs.brightway.dev/en/latest/content/theory/structure.html#database-is-a-subclass-of-datastore
-    Note that all activity data are optional.
-    When activities contain exchanges, some fields are required (input, type, amount)
-    Each exchange has a type: production, substitution, technosphere, or biosphere
-    AB does not yet support 'substitution'. Other exchange types are shown in separate columns on this interface
-    Required and other common exchange data fields are hardcoded as column headers in these tables
-    More detail available at: https://docs.brightway.dev/en/latest/content/theory/structure.html#exchange-data-format
-    The technosphere products (first table) of the visible activity are consumed by other activities downstream
-    The final table of this tab lists these 'Downstream Consumers'
-    """
-
-    def __init__(self, key: tuple, read_only=True, parent=None):
-        super(ActivityTab, self).__init__(parent)
-        self.read_only = read_only
-        self.db_read_only = project_settings.db_is_readonly(db_name=key[0])
-        self.key = key
-        self.db_name = key[0]
-        self.activity = bd.get_activity(key)
-        self.database = bd.Database(self.db_name)
-
-        # Edit Activity checkbox
-        self.checkbox_edit_act = QtWidgets.QCheckBox("Edit Activity")
-        self.checkbox_edit_act.setChecked(not self.read_only)
-        self.checkbox_edit_act.toggled.connect(lambda checked: self.act_read_only_changed(not checked))
-
-        # Activity Description
-        self.activity_description = SignalledPlainTextEdit(
-            key=key,
-            field="comment",
-            parent=self,
-        )
-
-        # Activity Description checkbox
-        self.checkbox_activity_description = QtWidgets.QCheckBox(
-            "Description", parent=self
-        )
-        self.checkbox_activity_description.clicked.connect(
-            self.toggle_activity_description_visibility
-        )
-        self.checkbox_activity_description.setChecked(not self.read_only)
-        self.checkbox_activity_description.setToolTip(
-            "Show/hide the activity description"
-        )
-        self.toggle_activity_description_visibility()
-
-        # Reveal/hide uncertainty columns
-        self.checkbox_uncertainty = QtWidgets.QCheckBox("Uncertainty")
-        self.checkbox_uncertainty.setToolTip("Show/hide the uncertainty columns")
-        self.checkbox_uncertainty.setChecked(False)
-        self.checkbox_uncertainty.toggled.connect(self.show_exchange_uncertainty)
-
-        # Reveal/hide exchange comment columns
-        self.checkbox_comment = QtWidgets.QCheckBox("Comments")
-        self.checkbox_comment.setToolTip("Show/hide the comment column")
-        self.checkbox_comment.setChecked(False)
-        self.checkbox_comment.toggled.connect(self.show_comments)
-
-        # Properties button
-        properties = QtWidgets.QPushButton("Properties")
-        properties.clicked.connect(self.open_property_editor)
-        properties.setToolTip("Show the properties dialog")
-
-        # Tags button
-        self.tags_button = QtWidgets.QPushButton("Tags")
-        self.tags_button.clicked.connect(self.open_tag_editor)
-        self.tags_button.setToolTip("Show the tags dialog")
-
-        # Toolbar Layout
-        toolbar = QtWidgets.QToolBar()
-        self.graph_action = toolbar.addAction(
-            qicons.graph_explorer, "Show in Graph Explorer", self.open_graph
-        )
-        toolbar.addWidget(self.checkbox_edit_act)
-        toolbar.addWidget(self.checkbox_activity_description)
-        toolbar.addWidget(self.checkbox_uncertainty)
-        toolbar.addWidget(self.checkbox_comment)
-        toolbar.addWidget(properties)
-        toolbar.addWidget(self.tags_button)
-
-        # Activity information
-        # this contains: activity name, location, database
-        self.activity_data_grid = ActivityDataGrid(
-            read_only=self.read_only, parent=self
-        )
-        self.db_read_only_changed(db_name=self.db_name, db_read_only=self.db_read_only)
-
-        # Exchange tables
-        self.production = ProductExchangeTable(self)
-        self.technosphere = TechnosphereExchangeTable(self)
-        self.biosphere = BiosphereExchangeTable(self)
-        self.downstream = DownstreamExchangeTable(self)
-
-        self.exchange_groups = [
-            DetailsGroupBox("Outputs:", self.production),
-            DetailsGroupBox("Inputs:", self.technosphere),
-            DetailsGroupBox("Biosphere Flows:", self.biosphere),
-            DetailsGroupBox("Downstream Consumers:", self.downstream),
-        ]
-        self.exchange_groups[-1].setChecked(False)  # hide Downstream table by default
-
-        self.group_splitter = QtWidgets.QSplitter(QtCore.Qt.Vertical)
-        for group in self.exchange_groups:
-            self.group_splitter.addWidget(group)
-        if state := ab_settings.settings.get("activity_table_layout", None):
-            self.group_splitter.restoreState(bytearray.fromhex(state))
-        self.group_splitter.setChildrenCollapsible(False)
-
-        # Full layout
-        layout = QtWidgets.QVBoxLayout()
-        layout.setContentsMargins(10, 10, 4, 1)
-        layout.setAlignment(QtCore.Qt.AlignTop)
-
-        layout.addWidget(toolbar)
-        layout.addWidget(self.activity_data_grid)
-        layout.addWidget(self.activity_description)
-        layout.addWidget(self.group_splitter)
-
-        self.setLayout(layout)
-
-        self.exchange_tables_read_only_changed()
-        self.populate()
-        self.update_tooltips()
-        self.update_style()
-        self.connect_signals()
-
-        # Make the activity tab editable in case it's new
-        if not self.read_only:
-            self.act_read_only_changed(False)
-
-    def connect_signals(self):
-        signals.database_read_only_changed.connect(self.db_read_only_changed)
-
-        signals.node.deleted.connect(self.on_node_deleted)
-        signals.database.deleted.connect(self.on_database_deleted)
-
-        signals.node.changed.connect(self.populate)
-        signals.edge.changed.connect(self.populate)
-        signals.edge.deleted.connect(self.populate)
-        signals.parameter.recalculated.connect(self.populate)
-
-        self.group_splitter.splitterMoved.connect(self.save_splitter_state)
-
-    def on_node_deleted(self, node):
-        if node.id == self.activity.id:
-            self.deleteLater()
-
-    def on_database_deleted(self, name):
-        if name == self.activity["database"]:
-            self.deleteLater()
-
-    @Slot(name="openGraph")
-    def open_graph(self) -> None:
-        signals.open_activity_graph_tab.emit(self.key)
-
-    @Slot(name="populatePage")
-    def populate(self) -> None:
-        """Populate the various tables and boxes within the Activity Detail tab"""
-        if self.db_name in bd.databases:
-            # Avoid a weird signal interaction in the tests
-            try:
-                self.activity = bd.get_activity(self.key)  # Refresh activity.
-            except DoesNotExist:
-                signals.close_activity_tab.emit(self.key)
-                return
-        self.populate_description_box()
-
-        # update the object name to be the activity name
-        self.setObjectName(self.activity["name"])
-
-        # fill in the values of the ActivityTab widgets, excluding the ActivityDataGrid which is populated separately
-
-        production = self.activity.production()
-        technosphere = self.activity.technosphere()
-
-        inputs = [x for x in production if x.amount >= 0] + [x for x in technosphere if x.amount < 0]
-        outputs = [x for x in production if x.amount < 0] + [x for x in technosphere if x.amount >= 0]
-
-        self.production.model.load(inputs)
-        self.technosphere.model.load(outputs)
-        self.biosphere.model.load(self.activity.biosphere())
-        self.downstream.model.load(self.activity.upstream())
-
-        self.show_exchange_uncertainty(self.checkbox_uncertainty.isChecked())
-        self.show_comments(self.checkbox_comment.isChecked())
-        self.activity_data_grid.populate()
-
-    def populate_description_box(self):
-        """Populate the activity description."""
-        self.activity_description.refresh_text(self.activity.get("comment", ""))
-        self.activity_description.setReadOnly(self.read_only)
-
-    @Slot(name="toggleDescription")
-    def toggle_activity_description_visibility(self) -> None:
-        """Show only if checkbox is checked."""
-        self.activity_description.setVisible(
-            self.checkbox_activity_description.isChecked()
-        )
-
-    @Slot(bool, name="toggleUncertaintyColumns")
-    def show_exchange_uncertainty(self, toggled: bool) -> None:
-        self.production.show_uncertainty(toggled)
-        self.technosphere.show_uncertainty(toggled)
-        self.biosphere.show_uncertainty(toggled)
-
-    @Slot(bool, name="toggleCommentColumn")
-    def show_comments(self, toggled: bool) -> None:
-        self.production.show_comments(toggled)
-        self.technosphere.show_comments(toggled)
-        self.biosphere.show_comments(toggled)
-
-    @Slot(bool, name="toggleReadOnly")
-    def act_read_only_changed(self, read_only: bool) -> None:
-        """When read_only=False specific data fields in the tables below become user-editable
-        When read_only=True these same fields become read-only"""
-        self.read_only = read_only
-        self.activity_description.setReadOnly(self.read_only)
-
-        if (
-            not self.read_only
-        ):  # update unique locations, units, etc. for editing (metadata)
-            signals.edit_activity.emit(self.db_name)
-
-        self.activity_data_grid.set_activity_fields_read_only(read_only=self.read_only)
-        self.activity_data_grid.populate_database_combo()
-        self.exchange_tables_read_only_changed()
-
-        self.update_tooltips()
-        self.update_style()
-
-    def exchange_tables_read_only_changed(self) -> None:
-        """The user should not be able to edit the exchange tables when read_only
-
-        Read-only flag set to True in models to prevent DoubleClick-selection editing.
-        DragDropMode set to NoDragDrop prevents exchanges dropped on the table to add.
-        """
-        for table in [
-            self.production,
-            self.technosphere,
-            self.biosphere,
-            self.downstream,
-        ]:
-            table.set_read_only_flag(self.read_only)
-
-    @Slot(str, bool, name="dbReadOnlyToggle")
-    def db_read_only_changed(self, db_name: str, db_read_only: bool) -> None:
-        """If database of open activity is set to read-only, the read-only checkbox cannot now be unchecked by user"""
-        if db_name == self.db_name:
-            self.db_read_only = db_read_only
-
-            # if activity was editable, but now the database is read-only, read_only state must be changed to false.
-            if not self.read_only and self.db_read_only:
-                self.checkbox_edit_act.setChecked(False)
-                self.act_read_only_changed(read_only=True)
-
-            # update checkbox to greyed-out or not
-            read_only_process = self.activity.get("type") == "readonly_process"
-            self.checkbox_edit_act.setEnabled(not self.db_read_only and not read_only_process)
-            self.update_tooltips()
-
-        else:  # on read-only state change for a database different to the open activity...
-            # update values in database list to ensure activity cannot be duplicated to read-only db
-            self.activity_data_grid.populate_database_combo()
-
-    def update_tooltips(self) -> None:
-        if self.db_read_only:
-            self.checkbox_edit_act.setToolTip(
-                "The database this activity belongs to is read-only."
-                " Enable database editing with checkbox in databases list"
-            )
-        else:
-            if self.read_only:
-                self.checkbox_edit_act.setToolTip(
-                    "Click to enable editing. Edits are saved automatically"
-                )
-            else:
-                self.checkbox_edit_act.setToolTip(
-                    "Click to prevent further edits. Edits are saved automatically"
-                )
-
-    def update_style(self) -> None:
-        # pass
-        if self.read_only:
-            self.setStyleSheet(style_activity_tab.style_sheet_read_only)
-        else:
-            self.setStyleSheet(style_activity_tab.style_sheet_editable)
-
-    def save_splitter_state(self):
-        ab_settings.settings["activity_table_layout"] = bytearray(
-            self.group_splitter.saveState()
-        ).hex()
-        ab_settings.write_settings()
-
-    def open_property_editor(self):
-        """Opens the property editor for the current """
-        # Do not save the changes if nothing changed
-        if PropertyEditor.edit_properties(self.activity, self.read_only, self):
-            self.activity.save()
-            # Properties changed, redo allocations, the values might have changed
-            MultifunctionalProcessRedoAllocation.run(self.activity)
-
-    def open_tag_editor(self):
-        """Opens the tag editor for the current"""
-        # Do not save the changes if nothing changed
-        if TagEditor.edit(self.activity, self.read_only, self):
-            self.activity.save()
